read mem info

+++ kernel/ukernio.h
@@ -7,7 +7,7 @@
 #define COLOR_BACK_BLACK 0x0000
 
 void print(char *);
-void printint(int n);
+void printint(uint64_t n);
 void newline();
 
 #endif
